+++ src/main/java/com.accenture.flowershop/be/business/UserBusinessServiceImpl.java
@@ -9,6 +9,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 @Service("userBusinessService")
@@ -53,7 +54,7 @@ public class UserBusinessServiceImpl implements UserBusinessService{
     }
 
     @Override
-    public Customer register(String login, String password, String surname, String name, String patronymic, String address, double cashBalance, int discount, UserShop userRole) {
+    public Customer register(String login, String password, String surname, String name, String patronymic, String address, BigDecimal cashBalance, BigDecimal discount, UserShop userRole) {
 
         Customer customer = new Customer(login, password, surname, name, patronymic, address, cashBalance, discount, userRole);
